+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Patch potential aborts in the vendored Swift demangler. ([#1054](https://github.com/getsentry/symbolic/pull/1054), [#1059](https://github.com/getsentry/symbolic/pull/1059), [#1062](https://github.com/getsentry/symbolic/pull/1062))
 - Do not allocate based on untrusted input in unreal parser. ([#1055](https://github.com/getsentry/symbolic/pull/1055))
 - Ensure unwind sections contain any unwind information. ([#1058](https://github.com/getsentry/symbolic/pull/1058))
+- Apply ELF relocations to debug sections in relocatable objects, fixing zero-valued file/directory names read from unlinked `.o` files. ([#1060](https://github.com/getsentry/symbolic/pull/1060))
 
 **Dependencies**